frontend/requirements.txt
streamlit
pydantic
google-api-python-client
google-auth
